issue #9: support `fuzzy` translations
When polib `merge` a new POT into the old PO it turns modified strings into
obsolete (commented out in the end of the PO). If the changes in the msgid are
small it is desirable to keep the translations as `fuzzy` so one can manually
sort out the changes on a editor (PoEdit).

This patch tries to reuse the obsolete translations. If there is a obsolete
translation `msgstr` and the msgids (old and new) are close enough (>65%) the
msgstr is reused and marked as 'fuzzy'.

A simple Levenshtein function is used to compute the distance between msgids.