Support no obsolete option (#65)
* tests: Sanity check line-width option

Provide a unit test to validate the usage of the `--line-width` option.

Signed-off-by: James Knight <james.d.knight@live.com>

* Adjust catalog commands to support forwarding options to babel

The following tweaks the catalog wrapper commands used to call babel's
`pofile`/`mofile` functions to support forwarding any keyword argument.
This can provide flexibility to callers which may wish to utilize
additional options provided by babel, without needing to add an explicit
option in the catalog wrapper calls.

A compatibility adjustment was added to `dump_po` to help map an
original argument `line_width` to babel's `width` argument. This is to
ensure callers which still use the `line_width` argument can
appropriately configure the babel call to impose the provided width.

Signed-off-by: James Knight <james.d.knight@live.com>

* Correct docstring spelling

Signed-off-by: James Knight <james.d.knight@live.com>

* Provide no-obsolete option

Adds support for a `--no-obsolete` option to allow the automatic
exclusion of obsolete messages (`#~ `) from updated sources. This
utilizes babel's `ignore_obsolete` option to not produce obsolete in a
final pot file.

The `--no-obsolete` argument was chosen to mimic the same argument
naming as seen in `msgattrib` [1].

[1]: https://www.gnu.org/software/gettext/manual/html_node/msgattrib-Invocation.html

Signed-off-by: James Knight <james.d.knight@live.com>

* tests: Validate newly added no-obsolete option

Signed-off-by: James Knight <james.d.knight@live.com>

---------

Signed-off-by: James Knight <james.d.knight@live.com>